Added the Ethan_Birb Model with animations.
Made the hybrid model of ethan and bird in Houdini, then uploaded to Mixamo to use their AutoRig to rig animations. We can now use Mixamo Animations to setup our own animation controller, and map different animations as per our requirement.
Issue: The tail feather of the bird distorts with the walk animation, will need to fix that in future iterations.